// front/src/services/types/api-schema.ts
export interface components {
    schemas: {
        UpdateProfileRequest: {
            name?: string;
            /** Format: date */
            birthDate?: string;
            address?: string;
            /** Format: email */
            email?: string;
            currentPassword?: string;
            newPassword?: string;
        };
        UserProfileResponse: {
            /** Format: uuid */
            id?: string;
            name?: string;
            email?: string;
            /** @enum {string} */
            role?: "Admin" | "Professor" | "Aluno";
            /** @enum {string} */
            status?: "Active" | "Inactive";
            /** Format: date */
            birthDate?: string;
            address?: string;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            updatedAt?: string;
        };
// ...
        LoginResponse: {
            accesToken?: string;
            refreshToken?: string;
        };
        RefreshTokenRequest: {
            refreshToken: string;
        };
        RefreshTokenResponse: {
            accessToken?: string;
            refreshToken?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}